+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- fix memory identity provider seeding
+
 ### Security
 
 ## [0.26.0] - 2024-05-29